empty_blockの実装
